pyproject.toml
[project]
name = "gabrr-budget-api"
version = "0.1.0"
description = "Parse financial documents (CSV/PDF) into normalized transactions using AI agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.0",
    "docling>=2.67.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "google-adk>=1.0.0",
    "litellm>=1.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
